# HospitalApp/Client/swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "API",
      "version": "1.0"
    },
    "servers":[
        {
            "url": "https://localhost:7067/"
        }
    ],
    "paths": {
      "/api/Auth/register": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/Auth/login": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/Auth/RefreshToken": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRequestDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRequestDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/Test": {
        "get": {
          "tags": [
            "Test"
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/WeatherForecast": {
        "get": {
          "tags": [
            "WeatherForecast"
          ],
          "operationId": "GetWeatherForecast",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/WeatherForecast"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/WeatherForecast"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/WeatherForecast"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            }
          },
          "security": [
            {
              "oauth2": [ ]
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "TokenRequestDto": {
          "required": [
            "refreshToken",
            "token"
          ],
          "type": "object",
          "properties": {
            "token": {
              "type": "string"
            },
            "refreshToken": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "UserDto": {
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "format": "email"
            },
            "password": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "WeatherForecast": {
          "type": "object",
          "properties": {
            "date": {
              "type": "string",
              "format": "date-time"
            },
            "temperatureC": {
              "type": "integer",
              "format": "int32"
            },
            "temperatureF": {
              "type": "integer",
              "format": "int32",
              "readOnly": true
            },
            "summary": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      },
      "securitySchemes": {
        "oauth2": {
          "type": "apiKey",
          "description": "Standard Authorization header using the Bearer scheme (\"bearer {token}\")",
          "name": "Authorization",
          "in": "header"
        }
      }
    }
  }